Move MCP frontend into the nvidia-nat-mcp package
