+++ packages/nvidia_nat_mcp/src/nat/plugins/mcp/server/memory_profiler.py
