+++ packages/nvidia_nat_mcp/src/nat/plugins/mcp/server/register_frontend.py
@@ -13,8 +13,15 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-# flake8: noqa
-# isort:skip_file
+from collections.abc import AsyncIterator
 
-from .console import register as console_register
-from .fastapi import register as fastapi_register
+from nat.cli.register_workflow import register_front_end
+from nat.data_models.config import Config
+from nat.plugins.mcp.server.front_end_config import MCPFrontEndConfig
+
+
+@register_front_end(config_type=MCPFrontEndConfig)
+async def register_mcp_front_end(config: MCPFrontEndConfig, full_config: Config) -> AsyncIterator:
+    from nat.plugins.mcp.server.front_end_plugin import MCPFrontEndPlugin
+
+    yield MCPFrontEndPlugin(full_config=full_config)
