+++ tests/nat/mcp/client/test_mcp_client_impl.py
@@ -21,12 +21,12 @@ from unittest.mock import patch
 from pydantic import BaseModel
 
 from nat.builder.workflow_builder import WorkflowBuilder
-from nat.plugins.mcp.client_base import MCPBaseClient
-from nat.plugins.mcp.client_config import MCPClientConfig
-from nat.plugins.mcp.client_config import MCPServerConfig
-from nat.plugins.mcp.client_config import MCPToolOverrideConfig
-from nat.plugins.mcp.client_impl import mcp_apply_tool_alias_and_description
-from nat.plugins.mcp.client_impl import mcp_client_function_group
+from nat.plugins.mcp.client.client_base import MCPBaseClient
+from nat.plugins.mcp.client.client_config import MCPClientConfig
+from nat.plugins.mcp.client.client_config import MCPServerConfig
+from nat.plugins.mcp.client.client_config import MCPToolOverrideConfig
+from nat.plugins.mcp.client.client_impl import mcp_apply_tool_alias_and_description
+from nat.plugins.mcp.client.client_impl import mcp_client_function_group
 
 
 class _InputSchema(BaseModel):
